+++ packages/editor/editor-core/src/utils/input-rules.ts
@@ -15,8 +15,8 @@ export function defaultInputRuleHandler(
     // Skip any input rule inside code
     // https://product-fabric.atlassian.net/wiki/spaces/E/pages/37945345/Editor+content+feature+rules#Editorcontent/featurerules-Rawtextblocks
     const unsupportedMarks = isBlockNodeRule
-      ? hasMarksNotSupportedByBlocks(state, start, end)
-      : hasCodeMark(state, start, end);
+      ? hasUnsupportedMarkForBlockInputRule(state, start, end)
+      : hasUnsupportedMarkForInputRule(state, start, end);
     if (state.selection.$from.parent.type.spec.code || unsupportedMarks) {
       return;
     }
@@ -49,7 +49,7 @@ export const uuid = () =>
     return (c === 'x' ? r : (r & 0x3) | 0x8).toString(16);
   });
 
-const hasMarksNotSupportedByBlocks = (
+const hasUnsupportedMarkForBlockInputRule = (
   state: EditorState,
   start: number,
   end: number,
@@ -57,7 +57,9 @@ const hasMarksNotSupportedByBlocks = (
   const { doc, schema: { marks } } = state;
   let unsupportedMarksPresent = false;
   const isUnsupportedMark = node =>
-    node.type === marks.code || node.type === marks.link;
+    node.type === marks.code ||
+    node.type === marks.link ||
+    node.type === marks.mentionQuery;
   doc.nodesBetween(start, end, node => {
     unsupportedMarksPresent =
       unsupportedMarksPresent ||
@@ -66,10 +68,15 @@ const hasMarksNotSupportedByBlocks = (
   return unsupportedMarksPresent;
 };
 
-const hasCodeMark = (state: EditorState, start: number, end: number) => {
+const hasUnsupportedMarkForInputRule = (
+  state: EditorState,
+  start: number,
+  end: number,
+) => {
   const { doc, schema: { marks } } = state;
   let unsupportedMarksPresent = false;
-  const isCodemark = node => node.type === marks.code;
+  const isCodemark = node =>
+    node.type === marks.code || node.type === marks.mentionQuery;
   doc.nodesBetween(start, end, node => {
     unsupportedMarksPresent =
       unsupportedMarksPresent || node.marks.filter(isCodemark).length > 0;