+++ website/src/pages/Package/ChangelogModal.js
@@ -42,9 +42,6 @@ const FieldWrapper = styled.div`
 `;
 const LogWrapper = styled.div`
   margin-top: 2em;
-  p {
-    display: none;
-  }
 `;
 
 // ==============================