+++ packages/core/field-text/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @atlaskit/field-text
 
+## 6.0.1
+- [patch] Update readme's [223cd67](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/223cd67)
+- [patch] Updated dependencies [223cd67](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/223cd67)
+  - @atlaskit/field-base@10.0.1
+  - @atlaskit/button@8.0.1
+  - @atlaskit/theme@4.0.1
+  - @atlaskit/docs@4.0.1
+
+## 6.0.0
+- [major] makes styled-components a peer dependency and upgrades version range from 1.4.6 - 3 to ^3.2.6 [1e80619](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/1e80619)
+- [patch] Updated dependencies [1e80619](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/1e80619)
+  - @atlaskit/field-base@10.0.0
+  - @atlaskit/button@8.0.0
+  - @atlaskit/theme@4.0.0
+  - @atlaskit/docs@4.0.0
+
 ## 5.1.1
 - [patch] Correct Flow type for FieldTextProps:onChange [adb2ce0](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/adb2ce0)
 - [none] Updated dependencies [adb2ce0](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/adb2ce0)